+++ crates/tiny-loop/src/tool.rs
@@ -25,9 +25,11 @@ pub trait Tool {
     async fn call_batch(&self, args: Vec<ToolCall>) -> Vec<Message> {
         join_all(
             args.into_iter()
-                .map(async |call| Message::Tool {
-                    tool_call_id: call.id,
-                    content: self.call(call.function.arguments).await,
+                .map(async |call| {
+                    Message::Tool(crate::types::ToolMessage {
+                        tool_call_id: call.id,
+                        content: self.call(call.function.arguments).await,
+                    })
                 })
                 .collect::<Vec<_>>(),
         )